Refactor Stripe integration
