+++ api/db/schema.rb
@@ -11,6 +11,7 @@
 # It's strongly recommended that you check this file into your version control system.
 
 ActiveRecord::Schema.define(version: 20170711041404) do
+
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
