+++ frontend/src/components/index.js
@@ -4,6 +4,7 @@ export { default as LoadingSpinner } from './LoadingSpinner/LoadingSpinner'
 export { default as Card } from './Card/Card'
 export { default as Cloud9SetupForm } from './Cloud9SetupForm/Cloud9SetupForm'
 export { default as Container } from './Container/Container'
+export { default as DonationForm } from './DonationForm/DonationForm'
 export { default as Emoji } from './Emoji/Emoji'
 export { default as Field } from './Field/Field'
 export { default as Header } from './Header/Header'
